Compute 'result' modulo n0*n1*n2.
Turns out the instructions were wrong, and you do need to do this. The algorithm given computes some number that's equivalent to m^3 in the expanded n0n1n2 domain, but cube-rooting that will not necessarily get you m (for some reason. I might be confused about the math here.)